реализациz интерфейса выборки из БД

+++ src/main/java/ru/sumenkov/dspsql/repository/SearchRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface SearchRepository {
     List<BuyerModel> searchBuyer(String lastname);
     List<BuyerModel> searchProduct(String name, int amount);
-    List<BuyerModel> searchExpenses(double min, double max);
+    List<BuyerModel> searchExpenses(int min, int max);
     List<BuyerModel> searchBadCustomers(int amount);
 }
